% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{roseRF_plm}
\alias{roseRF_plm}
\title{ROSE random forest estimator for the partially linear model}
\usage{
roseRF_plm(
  y_formula,
  y_learner,
  y_pars = list(),
  x_formula,
  x_learner,
  x_pars = list(),
  M1_formula = x_formula,
  M1_learner = x_learner,
  M1_pars = x_pars,
  M2_formula = NA,
  M2_learner = NA,
  M2_pars = list(),
  M3_formula = NA,
  M3_learner = NA,
  M3_pars = list(),
  M4_formula = NA,
  M4_learner = NA,
  M4_pars = list(),
  M5_formula = NA,
  M5_learner = NA,
  M5_pars = list(),
  data,
  K = 5,
  S = 1,
  max.depth = 10,
  num.trees = 500,
  min.node.size = max(10, ceiling(0.01 * (K - 1)/K * nrow(data))),
  replace = TRUE,
  sample.fraction = 0.8
)
}
\arguments{
\item{y_formula}{a two-sided formula object describing the model for \eqn{E[Y|Z]}.}

\item{y_learner}{a string specifying the regression method to fit the regression of \eqn{Y} on \eqn{Z} as given by \code{y_formula} (e.g. \code{randomforest, xgboost, neuralnet, gam}).}

\item{y_pars}{a list containing hyperparameters for the \code{y_learner} chosen. Default is an empty list, which performs hyperparameter tuning.}

\item{x_formula}{a two-sided formula object describing the model for \eqn{E[X|Z]}.}

\item{x_learner}{a string specifying the regression method to fit the regression of \eqn{X} on \eqn{Z} as given by \code{x_formula} (e.g. \code{randomforest, xgboost, neuralnet, gam}).}

\item{x_pars}{a list containing hyperparameters for the \code{y_learner} chosen. Default is an empty list, which performs hyperparameter tuning.}

\item{M1_formula}{a two-sided formula object for the model \eqn{E[M_1(X,Z)|Z]}. Default is \eqn{M_1(X,Z)=X}.}

\item{M1_learner}{a string specifying the regression method for \eqn{E[M_1(X,Z)|Z]} estimation.}

\item{M1_pars}{a list containing hyperparameters for the \code{M1_learner} chosen.}

\item{M2_formula}{a two-sided formula object for the model \eqn{E[M_2(X,Z)|Z]}. Default is no formula / regression (i.e. \eqn{J=1})}

\item{M2_learner}{a string specifying the regression method for \eqn{E[M_2(X,Z)|Z]} estimation.}

\item{M2_pars}{a list containing hyperparameters for the \code{M2_learner} chosen.}

\item{M3_formula}{a two-sided formula object for the model \eqn{E[M_3(X,Z)|Z]}. Default is no formula / regression (i.e. \eqn{J=1}).}

\item{M3_learner}{a string specifying the regression method for \eqn{E[M_3(X,Z)|Z]} estimation.}

\item{M3_pars}{a list containing hyperparameters for the \code{M3_learner} chosen.}

\item{M4_formula}{a two-sided formula object for the model \eqn{E[M_4(X,Z)|Z]}. Default is no formula / regression (i.e. \eqn{J=1})}

\item{M4_learner}{a string specifying the regression method for \eqn{E[M_4(X,Z)|Z]} estimation.}

\item{M4_pars}{a list containing hyperparameters for the \code{M4_learner} chosen.}

\item{M5_formula}{a two-sided formula object for the model \eqn{E[M_5(X,Z)|Z]}. Default is no formula / regression (i.e. \eqn{J=1})}

\item{M5_learner}{a string specifying the regression method for \eqn{E[M_5(X,Z)|Z]} estimation.}

\item{M5_pars}{a list containing hyperparameters for the \code{M5_learner} chosen.}

\item{data}{a data frame containing the variables for the partially linear model.}

\item{K}{the number of folds used for \eqn{K}-fold cross-fitting. Default is 5.}

\item{S}{the number of repeats to mitigate the randomness in the estimator on the sample splits used for \eqn{K}-fold cross-fitting. Default is 5.}

\item{max.depth}{Maximum depth parameter used for ROSE random forests. Default is 5.}

\item{num.trees}{Number of trees used for a single ROSE random forest. Default is 50.}

\item{min.node.size}{Minimum node size of a leaf in each tree. Default is \code{max(10,ceiling(0.01 (K-1)/K nrow(data)))}.}

\item{replace}{Whether sampling for a single random tree are performed with (bootstrap) or without replacement. Default is \code{TRUE} (i.e. bootstrap).}

\item{sample.fraction}{Proportion of data used for each random tree. Default is 0.8.}
}
\value{
A list containing:
  \describe{
    \item{\code{theta}}{The estimator of \eqn{\theta_0}.}
    \item{\code{stderror}}{Huber robust estimate of the standard error of the \eqn{\theta_0}-estimator.}
    \item{\code{coefficients}}{Table of \eqn{\theta_0} coefficient estimator, standard error, z-value and p-value.}
  }
}
\description{
Estimates the parameter of interest \eqn{\theta_0} in the partially linear model
\deqn{E[Y|X,Z] = X\theta_0 + f_0(Z),} which can be reposed in terms of
the `nuisance functions' \eqn{(E[Y|X], E[X|Z])} as \deqn{E[Y|X,Z]-E[Y|Z] = (X-E[X|Z])\theta_0.}
}
\details{
The estimator of interest \eqn{\theta_0} solves the estimating equation
\deqn{\sum_{i}\psi(Y_i,X_i,Z_i; \theta,\hat{\eta}(Z),\hat{w}(Z)) = 0,}
\deqn{\psi(Y,X,Z;\theta,\eta_0,w) := \Big(\sum_{j=1}^J w_j(Z) \big( M_j(X,Z) - E[M_j(X,Z)|Z] \big) \Big) \Big( \big(Y-E[Y|Z]\big)-\big(X-E[X|Z]\big)\theta \Big),}
\deqn{\eta_0 := \big(E[Y|Z=\cdot], E[X|Z=\cdot]\big),}
where \eqn{M_1(X,Z),\ldots,M_J(X,Z)} denotes user-chosen functions of \eqn{(X,Z)}
and \eqn{w(Z)=\big(w_1(Z),\ldots,w_J(Z)\big)} denotes weights estimated via ROSE random forests.
The recommended default takes \eqn{J=1} and \eqn{M_1(X,Z)=X}; if taking \eqn{J\geq 2} we recommend care
in checking the applicability and appropriateness of any additional user-chosen
regression tasks.

The parameter of interest \eqn{\theta_0} is estimated using a DML2 / \eqn{K}-fold cross-fitting
framework, to allow for arbitrary (\eqn{n^{1/4}}-consistent) learners for \eqn{\hat{\eta}} i.e. solving
the estimating equation
\deqn{\sum_{k}\sum_{I_k}\psi(Y_i,X_i,Z_i; \theta,\hat{\eta}^{(k)}(Z),\hat{w}^{(k)}(Z)) = 0,}
where \eqn{I_1,\ldots,I_K} denotes a partition of the index set for the datapoints \eqn{(Y_i,X_i,Z_i)},
\eqn{\hat{\eta}^{(k)}} denotes an estimator for \eqn{\eta_0} trained on the data indexed by
\eqn{I_k^c}, and \eqn{\hat{w}^{(k)}} denotes a ROSE random forest (again trained on the data
indexed by \eqn{I_k^c}).
}
