% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.roseforest.R
\name{summary.roseforest}
\alias{summary.roseforest}
\title{Summary for a rose random forest fitted object}
\usage{
\method{summary}{roseforest}(object, ...)
}
\arguments{
\item{object}{a fitted \code{roseRF} object fitted by \code{roseRF_...}.}

\item{...}{additional arguments}
}
\value{
Prints summary output for \code{roseRF} object
}
\description{
Prints a \code{roseRF} object fitted by the functions \code{roseRF_...} in \code{roseRF}.
}
