% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.roseforest.R
\name{print.roseforest}
\alias{print.roseforest}
\title{Print for a rose random forest fitted object}
\usage{
\method{print}{roseforest}(x, ...)
}
\arguments{
\item{x}{a fitted \code{rose} object fitted by \code{rose_...}.}

\item{...}{additional arguments}
}
\description{
This is a method that prints a useful summary of aspects of a \code{rose} object fitted by the functions \code{rose_...} in \code{rose}.
}
