% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.roseforest.R
\name{summary.roseforest}
\alias{summary.roseforest}
\title{Summary for a rose random forest fitted object}
\usage{
\method{summary}{roseforest}(object, ...)
}
\arguments{
\item{object}{a fitted \code{rose} object fitted by \code{rose_...}.}

\item{...}{additional arguments}
}
\description{
Prints a \code{rose} object fitted by the functions \code{rose_...} in \code{rose}.
}
